+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog Capstone-Laravel-tugboat-dev-PreFinal α
 ### `Pre Final Repository Monday August 27,2018`
 
+### Consignee Contract Request Page Edited
+- [x] Added matrix table and custom matrix suggestion fields
+
 ### Contracts Module
 - [x] UI For Contract Page
 &nbsp;